# Improve performance/reduce allocs of bytespipe